+++ macros/alchemy/potion-potion-of-full-healing.js
@@ -4,11 +4,9 @@ const itemName = 'Potion of Full Healing';
 source: 
 icon: 
 */
-const version = 'v0.1';
+const version = 'v0.2';
 let tokenD=canvas.tokens.controlled[0];
 const myTitle = `Potion`;
-let message1 = ``;
-let message2 = itemName;
   
 if (tokenD===undefined) {
   ui.notifications.error("Please select a token."); // No Token is Selected
@@ -25,7 +23,7 @@ async function main() {
   
   //If token is max health if so, don't do anything
   if( sm.getWounds(tokenD) == 0 ) {
-    sm.styledChatMessage(itemName, '', `You are not injured.`)       
+    sm.styledChatMessageSimple(itemName, `You are not injured.`)       
     return;
   }
 
@@ -35,5 +33,5 @@ async function main() {
   // remove 1 current wound
   sm.applyWounds(tokenD, -sm.getWounds(tokenD));
 
-  sm.styledChatMessage(itemName, '', `${tokenD.actor.name} used a ${itemName} and cured ${sm.getWounds(tokenD)} wounds.`);
+  sm.styledChatMessageSimple(itemName, `${tokenD.actor.name} used a ${itemName} and cured ${sm.getWounds(tokenD)} wounds.`);
 }