+++ macros/alchemy/potion-potion-of-speed.js
@@ -1,12 +1,13 @@
-const version = 'v0.2';
+const version = 'v0.3';
 const itemName = 'Potion of Speed';
+const icon = 'icons/consumables/potions/bottle-conical-corked-purple.webp';
 
 const myActiveEffect = {
   changes: [
     {"key":"data.stats.speed.value", "mode":CONST.ACTIVE_EFFECT_MODES.ADD, "value":"2"},
     {"key":"data.stats.speed.runningDie", "mode":CONST.ACTIVE_EFFECT_MODES.ADD, "value":"2"}
   ],
-  icon: 'icons/consumables/potions/bottle-conical-corked-purple.webp',
+  icon: icon,
   label: itemName
 }  
 
@@ -16,8 +17,6 @@ icon:
 */
 let tokenD=canvas.tokens.controlled[0];
 const myTitle = `Potion`;
-let message1 = ``;
-let message2 = itemName;
   
 if (tokenD===undefined) {
   ui.notifications.error("Please select a token."); // No Token is Selected
@@ -35,8 +34,7 @@ async function main() {
   // Active Effect On?
   const temp = sm.getActiveEffect(tokenD, itemName);
   if( temp != undefined ) {
-    console.log(sm.getActiveEffect(tokenD, itemName));
-    sm.styledChatMessage(itemName, '', `You already used it.`)       
+    sm.styledChatMessageSimple(itemName, `You already used it.`, icon);       
     return;
   }
 
@@ -47,5 +45,5 @@ async function main() {
   sm.addActiveEffectToOwnedToken(tokenD, myActiveEffect); 
 
   // message
-  sm.styledChatMessage(itemName, '', `${tokenD.actor.name} used a ${itemName}.`);
+  sm.styledChatMessageSimple(itemName, `${tokenD.actor.name} used a ${itemName}.`, icon);
 }