+++ macros/alchemy/potion-potion-of-strong-healing.js
@@ -4,11 +4,8 @@ const itemName = 'Potion of Strong Healing';
 source: 
 icon: 
 */
-const version = 'v0.2';
+const version = 'v0.3';
 let tokenD=canvas.tokens.controlled[0];
-const myTitle = `Potion`;
-let message1 = ``;
-let message2 = itemName;
   
 if (tokenD===undefined) {
   ui.notifications.error("Please select a token."); // No Token is Selected
@@ -25,7 +22,7 @@ async function main() {
   
   //If token is max health if so, don't do anything
   if( sm.getWounds(tokenD) == 0 ) {
-    sm.styledChatMessage(itemName, '', `You are not injured.`)       
+    sm.styledChatMessageSimple(itemName, `You are not injured.`);       
     return;
   }
 
@@ -35,5 +32,5 @@ async function main() {
   // remove 1 current wound
   sm.applyWounds(tokenD, -2);
 
-  sm.styledChatMessage(itemName, '', `${tokenD.actor.name} used a ${itemName} and cured 2 wound.`);
+  sm.styledChatMessageSimple(itemName, `${tokenD.actor.name} used a ${itemName} and cured 2 wound.`);
 }